Remove all fail fast stuff #29
Now that we are pre-divvying there is currently no way to check for
done mid run. In order to bring this feature back we would need some
time during the life of a worker when it can be interrupted. Could be a
req rep for status to the sink when a test is finished. Until there is
more demand that isn't worth doing.