Continue even if Process.setpgrp fails
In CircleCI, calling Process.setpgrp raises an "Operation not permitted" error,
but other than that flatware works.

In this change, we allow flatware to continue when the error is raised.

If flatware is executed with --log, then it will print:
INFO -- : flatware sink continuing after: Process.setpgrp: Operation not permitted

At Carwow, we have been running flatware on CircleCI for about 1 month
(with a monkey-patched Process.setpgrp) and it works great!

It allowed Carwow to reduce CircleCI credits by half, and also our Docker Hub pulls.