Add option to distribute jobs by example instead of by file
The purpose of this change is to allow more optimal distribution when
the suite contains outlier files.

With enough cores on a CI run we noticed that the lower bound of the
duration of the full run was the runtime of the slowest file.

In order to be able to split by example, ExampleJobBuilder needs to load
all spec files and extract examples from them. As a side effect, this
unlocks more filtering options (such as by tag) which previously would
silently not work as expected.